fix: Make sure remove older BE container when it is exist

+++ rm-old-be-container.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
 
 id=$(docker ps -a -q -f "name=backend")
-docker container rm -f $id
+if [ -n $id ]; 
+then
+  docker container rm -f $id
